backend-usermanagement/requirements.txt
psycopg2-binary==2.9.9
django==5.0.1
django-cockroachdb==5.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
